Cleanup

+++ Psydekick/Source/Psydekick/Public/Psydekick_Visuals_2D.h
@@ -8,7 +8,7 @@
 
 #include "Psydekick_Visuals_2D.generated.h"
 
-DECLARE_DYNAMIC_DELEGATE_TwoParams(FChoiceMade, FString, choice, uint8, index);
+DECLARE_DYNAMIC_DELEGATE_TwoParams(FChoiceMade, FString, choice, int32, index);
 
 UCLASS()
 class PSYDEKICK_API UPsydekick_Visuals_2D : public UBlueprintFunctionLibrary
@@ -26,7 +26,7 @@ public:
 	static void ShowImage(const UObject* WorldContextObject, UTexture* Image);
 
 	UFUNCTION(BlueprintCallable, Category = "Psydekick|Visuals|2D", meta = (WorldContext = WorldContextObject))
-	static void GetChoice(const UObject* WorldContextObject, const FString Prompt, const TArray<FString> Options, const FChoiceMade &ChoiceMade);
+	static void GetChoice(const UObject* WorldContextObject, const FString Prompt, const TArray<FString> Options, const FChoiceMade& ChoiceMade);
 
 	static TSharedPtr<class SWidget> CurrentWidget;
 };
